agent_service/requirements.txt
google-adk>=1.30
google-genai>=1.10
fastapi>=0.115
uvicorn[standard]>=0.30
httpx>=0.27
pydantic>=2.7
python-dotenv>=1.0
